% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allSignalTTest.R
\name{allSignalTTests}
\alias{allSignalTTests}
\title{Calculate t-test-statistics for differences between week0 and surrounding weeks
of signal functions for all activity types and signals}
\usage{
allSignalTTests(
  allvar,
  stats,
  events,
  first = FALSE,
  dels = 0,
  weighted = FALSE,
  EVcol,
  sepPlots = NULL,
  inds,
  weeks = 1:15,
  week0 = 7,
  weekcorrect = 0,
  plotit = FALSE,
  devs,
  yranges = NULL,
  legend = TRUE
)
}
\arguments{
\item{allvar}{A list of \code{data.frames} with FOI activity data.
See description below for
more information about the content. The signal names should be used as
identifiers for each  \code{data.frame} in the list.}

\item{stats}{The statistical descriptors}

\item{events}{A vector with the events under \code{dat$EVENT_TYPE} that should be included}

\item{first}{Logical; TRUE if only the first event of the year of each
FOI should be included.}

\item{dels}{The minimum number of days (or a vector of minimum number of days) between two activities for
one of them to be included in the mean.}

\item{weighted}{Logical; TRUE if the mean should be weighted according
to the number of acquisitions used for the weekly mean.}

\item{EVcol}{The column giving the activity type}

\item{sepPlots}{List of groups of indexes that should be plotted for
a limited group of descriptors in separate devices (normally pdfs).
It can either be in a
format similar to \code{sepPlots = list(p1 = c("s1_coh6", "s1_coh6_VH"), 
                p2 = c("NDVI", "BSI"))}, which will plot to two extra devices,
two s1-signals to the first and 2 s2-indexes to the second, for the
default descriptor \code{P50}. Otherwise it can be given in the format
sepPlots = list(bands = list(p1 = c("s1_coh6", "s1_coh6_VH"),
p2 = c("NDVI", "BSI"), stats = c("P50", "MEAN"), which
plot to the same extra devices, but will also include the \code{MEAN} values.
The number of groups in \code{sepPlots} should be reflected by
the number of devices \code{devs}}

\item{inds}{Array with names of all indices, same order as used in evres}

\item{weeks}{The number of weeks used for the signal, taken as all values of \code{dat$WEEK} if missing}

\item{week0}{If a signal consists of \code{weeks}, \code{w0} describes the week to compare with}

\item{weekcorrect}{If the week numbers should be shifted, so that 0 will be one or more weeks after week0}

\item{plotit}{Logical; should plots be generated?}

\item{devs}{One or more devices for plotting the results of the function,
or one or more file names that will be opened as pdfs in the function.
The number of devices should reflect the number of groups in sepPlots + 1. Alternatively the base of one file name can be given, and the necessary
number of pdfs will be opened.  If only one device is given, all the plots will
go to this device. If none is given, all plots will go to the active device}

\item{yranges}{A list of plotting ranges for the y-axis, one for each statistical descriptor}

\item{legend}{If legend is TRUE or "YES", a legend will be added to each plot.
If legend = "BEFORE"),  there will be a separate page at the start of the
document with all legends instead of a legend on each plot.
The argument is not case sensitive.}
}
\description{
Calculate t-test-statistics for differences between week0 and surrounding weeks
of signal functions for all activity types and signals
}
\examples{
ddir = system.file("extdata", package = "SALMS")
odir = paste0(ddir, "/signals")

hdir = tempdir()
# Read all signals and find the names of all signals and indicators
allvar = readSignals(odir)
inds = names(allvar)
# Create three pdfs for t-test output
devs = list()
for (ipdf in 1:3) {pdf(paste0(hdir, "/ttest", ipdf, ".pdf")) ; devs[ipdf] = dev.cur()}

dfs = allSignalTTests(allvar, first = FALSE, dels = 0, weighted = FALSE,
                     inds = inds, dev = devs, weeks = 1:15, 
                     week0 = 7, weekcorrect = 0, plotit = TRUE, 
                     legend = "before") 
# Close the pdfs
for (ipdf in 1:3) dev.off(devs[ipdf])

}
\references{
{
\insertRef{Zielinski2022}{SALMS}
}
}
