% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bestSignals.R
\name{optimalSignal}
\alias{optimalSignal}
\title{Function that ranks the probabilities of different statistical descriptors of
a signal being above a threshold}
\usage{
optimalSignal(
  events,
  inds,
  evres,
  stats,
  week0 = 7,
  weeks = 1:15,
  selwks = 10:11
)
}
\arguments{
\item{events}{A vector with the events under \code{dat$EVENT_TYPE} that should be included}

\item{inds}{Array with names of all indices, same order as used in evres}

\item{evres}{A 5-dimensional array with means and probabilities, generated by functions in
the SALMS-package. See below for more details}

\item{stats}{The statistical descriptors}

\item{week0}{What should be the week of activity in the centred time series? This parameter  also defines the length of the time centred time series, which will be \code{2*week0 + 1}}

\item{weeks}{The number of weeks used for the signal, taken as all values of \code{dat$WEEK} if missing}

\item{selwks}{For which weeks should the descriptive statistics be shown}
}
\value{
The function will return a \code{data.frame} with a column for each descriptor
with ranking.
}
\description{
Function that ranks the probabilities of different statistical descriptors of
a signal being above a threshold
}
\details{
The function summarizes descriptors for all signals (ranks the best and worst)
for the selected
weeks (\code{selwks}). This selection is done based on the probability of the signal
being above zero for the selected weeks.
This can be used as an indication of which descriptor is the
most and least reliable. Additionally, it summarizes what can be regarded as optimal descriptors,
i.e., summarizing the three best descriptors for each signal. This is done, as it can sometimes
be minor differences between the best and the second best signals.

evres is a 5-dimensional array, with information about the mean changes
and probabilities of the change
being above a threshold. It can both be an input argument and the result of different functions in the
SALMS-package. If there is only one activity type, the first dimension can be dropped.
The dimensions are \code{c(events, inds, stats, evres-var , weeks)}. The content of the
array is defined by the different values of \code{evres-var}. These can be:
\enumerate{
\item{1 \code{mean(x)}}
\item{2 \code{cumsum(x)}}
\item{3 \code{std(x)}}
\item{4 \code{std(cumsum(x))}}
\item{5 \code{mean(x)*sign(mean(xx))} where \code{xx} is the signal change for the weeks \code{selwks}  (Making sure the average change after an activity is positive)}
\item{6 \code{cumsum(x)*sign(mean(xx))} where \code{xx} is the signal change for the weeks \code{selwks}  (Making sure the average change after an activity is positive)}
\item{7 \code{p0} - The probability of the change being larger than 0}
\item{8 \code{p01} - The probability of the change being later than a threshold}
\item{9 \code{p0c} - The probability of the cumulative change being larger than 0}
\item{10 \code{p01c} - The probability of the cumulative change being larger than a threshold}
\item{11 \code{max(po)} - The highest p0 of interesting weeks}
\item{12 \code{max(p01)} - The highest p01 of interesting weeks}
\item{13 The value of the change relative to highest}
\item{14 The value of the change relative to lowest}
\item{15 The value \code{p0} relative to the highest}
\item{16 The value \code{p0} relative to the lowest}
}
}
