% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeSeries.R
\name{plotTimeSeries}
\alias{plotTimeSeries}
\title{Plotting of signal time series}
\usage{
plotTimeSeries(
  FOIinfo,
  FOIs,
  stats,
  plotAllStats = TRUE,
  sepPlots = NULL,
  cols,
  obands,
  devs,
  tsdir,
  iyear,
  yranges,
  removeClouds = TRUE,
  cloudinds,
  addEvent = TRUE,
  addText = FALSE,
  legend = TRUE
)
}
\arguments{
\item{FOIinfo}{\code{\link{data.frame}} with information about activities and the corresponding FOIs. See details}

\item{FOIs}{an array of FOI IDs to be plotted}

\item{stats}{The statistical descriptors}

\item{plotAllStats}{Logical; if FALSE only the quantiles and the mean will be plotted}

\item{sepPlots}{List of groups of indexes that should be plotted for
a limited group of descriptors in separate devices (normally pdfs).
It can either be in a
format similar to \code{sepPlots = list(p1 = c("s1_coh6", "s1_coh6_VH"), 
                p2 = c("NDVI", "BSI"))}, which will plot to two extra devices,
two s1-signals to the first and 2 s2-indexes to the second, for the
default descriptor \code{P50}. Otherwise it can be given in the format
sepPlots = list(bands = list(p1 = c("s1_coh6", "s1_coh6_VH"),
p2 = c("NDVI", "BSI"), stats = c("P50", "MEAN"), which
plot to the same extra devices, but will also include the \code{MEAN} values.
The number of groups in \code{sepPlots} should be reflected by
the number of devices \code{devs}}

\item{cols}{array with colors}

\item{obands}{A list of output bands, grouped according to sbss, if missing, the default bands are used}

\item{devs}{One or more devices for plotting the results of the function,
or one or more file names that will be opened as pdfs in the function.
The number of devices should reflect the number of groups in sepPlots + 1.  If only one device is given, all the plots will
go to this device. If none is given, all plots will go to the active device}

\item{tsdir}{where to find the time series}

\item{iyear}{The year to plot. Can also be found from data, but will complain if there are more
than one year in the data set.}

\item{yranges}{A list of plotting ranges for the y-axis, one for each statistical descriptor}

\item{removeClouds}{Logical; Should clouds be removed}

\item{cloudinds}{Which indexes are regarded as clouds and can be discarded (for S2 images). Cloud filtering
can also be done externally, introducing a column VALID_OBS with 0 for invalid acquisitions}

\item{addEvent}{Should a vertical line be added for the events (from column EVENT_TYPE)}

\item{addText}{Should the line be accompanied with the event type itself}

\item{legend}{If legend is TRUE or "YES", a legend will be added to each plot.
If legend = "BEFORE"),  there will be a separate page at the start of the
document with all legends instead of a legend on each plot.
The argument is not case sensitive.}
}
\description{
Function plotting the time series of signals for each FOI, for the
different descriptors and bands
}
\details{
The function can take two output devices (or file names) for the plotting.
In the first device, the function will first produce a plot for each descriptor, grouping together the
bands for each sensor/index type (s1_coh6, S2, s1_bs, s2_idx, ...). Then it will produce plots for
each band, showing all the different descriptors in the same plot. This plot will typically have a
large number of pages

The function will produce considerably fewer pages for the second device, making it easier to
look through the output. The function will only make coherence plots, for p50/median.
}
\examples{
# Create a limited set of bands
bands = list(s1_coh6 = paste0("s1_coh6", c( "_VH_A",  "_VV_A")), 
    s2 = c("B02", "B03", "B04", "B05"), 
    s1_bs = paste0("s1_bs", c("_VH",  "_VH_A", "_VH_D")),
    s2_idx = c("NDVI", "NDWI", "NDPI", "OSAVI", "SAVI"), 
    s1_bsidx = c("RVI",  "CPRI"),
    s1_coh6rat = c("s1_coh6_VH_rAD", "s1_coh6_rADN", "s1_coh6_VV_rADN"),
    s1_bsrat = c("s1_bs_rAD", "s1_bs_rADN", "s1_bs_VH_rADN", "s1_bs_VV_rADN"))


# Directory names and file names have to be modified
tsdir = system.file("extdata", package = "SALMS")
FOIinfo = read.csv(paste0(tsdir, "/gtdir/FOIinfo.csv"))
tsdir = paste0(tsdir, "/tsdir")
hdir = tempdir() 
devs = list()
pdf(paste0(hdir, "/ts1.pdf"))
devs[[1]] = dev.cur()
pdf(paste0(hdir, "/ts2.pdf"))
devs[[2]] = dev.cur()

plotTimeSeries(FOIinfo,  devs = devs, tsdir = tsdir, obands = bands, iyear = 2020, 
     legend = "before") 
graphics.off()

}
\references{
{
\insertRef{Zielinski2022}{SALMS}
}
}
