% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSignals.R
\name{readSignals}
\alias{readSignals}
\title{Function to read activity centred signals from .csv-files and store them in a list}
\usage{
readSignals(odir, stats, zeroshift = FALSE, week0 = 7)
}
\arguments{
\item{odir}{The directory of the activity centred time series}

\item{stats}{The statistical descriptors}

\item{zeroshift}{Logical; whether the weekly centred time series should be shifted according to the value of week0}

\item{week0}{If a signal consists of \code{weeks}, \code{w0} describes the week to compare with}
}
\description{
Function to read activity centred signals from .csv-files and store them in a list
}
\details{
The function reads .csv files from the directory \\code{odir}
The files should have the name of the signal. The column names should
be what is typically produced from \\code{\\link{createSignals}}

\\code{week0} should only be set if also \\code{zeroshift = TRUE}.
}
