% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Correlations.R
\name{signalCorrelations}
\alias{signalCorrelations}
\title{Creates and plots correlations of the signals}
\usage{
signalCorrelations(
  allvar,
  events,
  inds,
  stats,
  custom = "ALL",
  mar = NULL,
  drop = TRUE,
  ...
)
}
\arguments{
\item{allvar}{A list of \code{data.frames} with FOI activity data.
See description below for
more information about the content. The signal names should be used as
identifiers for each  \code{data.frame} in the list.}

\item{events}{A vector with the events under \code{dat$EVENT_TYPE} that should be included}

\item{inds}{Array with names of all indices, same order as used in evres}

\item{stats}{The statistical descriptors}

\item{custom}{Either "ALL" or a list of different combinations of indexes
to be grouped in the plots}

\item{mar}{The margins, as defined in \code{\link{par}}. This parameter is set explicitly
in the function if missing, as the default value is not good for
the title.}

\item{drop}{Should the function only return an array if there is only
one activity type, instead of a list of length one}

\item{...}{Other parameters to \code{\link[corrplot]{corrplot}}}
}
\value{
The function will return the correlations as a multidimensional array,
or a list of arrays if there are more than one activity type in the
data set. The dimensions are length(inds), length(inds), length(descriptors)
}
\description{
Creates and plots correlations of the signals
}
\details{
The function produces different correlation plots for the individual
statistical descriptors of the data set. The plots can either be sent to
a file (typically multipage pdf) or to the screen, where next plots
are shown after hitting the return button.
}
\examples{
ddir = system.file("extdata", package = "SALMS")
odir = paste0(ddir, "/signals")
allvar = readSignals(odir)
allcorrs = signalCorrelations(allvar, stats = c("MIN", "P50"))
allcorrs = signalCorrelations(allvar, stats = c("MIN", "P50"),
                  custom = list(first = c("B04", "B08", "NDVI", "RVI"),
                                second = c("BSI", "GARI", "s1_bs_rAD", "s1_coh6")))

}
