% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signalMeans.R
\name{signalMean}
\alias{signalMean}
\title{Create means of signal functions for an activity type}
\usage{
signalMean(
  dat,
  stats,
  events,
  first = FALSE,
  dels = 0,
  weighted = FALSE,
  weeks,
  week0 = 7,
  weekcorrect = 0,
  plotit = FALSE,
  plotCumsum = FALSE,
  devs,
  sepPlots = NULL,
  zeroshift = TRUE,
  xlab = "weeks",
  ylab = "",
  main = "",
  yrange = NULL,
  FOIinfo,
  ind = ind,
  EVDcol,
  legend = TRUE,
  addNum = TRUE,
  iprint = 1
)
}
\arguments{
\item{dat}{A \code{data.frame} with FOI activity data. See description below for more information about the content}

\item{stats}{The statistical descriptors}

\item{events}{If only a certain type of events in \code{dat} should be used}

\item{first}{Logical; TRUE if only the first event of the year of each
FOI should be included.}

\item{dels}{The minimum number of days (or a vector of minimum number of days) between two activities for
one of them to be included in the mean.}

\item{weighted}{Logical; TRUE if the mean should be weighted according
to the number of acquisitions used for the weekly mean.}

\item{weeks}{The number of weeks used for the signal, taken as all values of
\code{dat$WEEK} if missing. Only necessary if \code{plotit = TRUE}}

\item{week0}{If a signal consists of \code{weeks}, \code{w0} describes the week to compare with}

\item{weekcorrect}{If the week numbers should be shifted, so that 0 will be one or more weeks after week0}

\item{plotit}{Should the function plot anything? Either "confidence", "extra"
or "both"/TRUE for all plots.}

\item{plotCumsum}{Should also the cumulative sum (before/after week0) be plotted?}

\item{devs}{One or more devices for plotting the results of the function,
or one or more file names that will be opened as pdfs in the function.
The number of devices should reflect the number of groups in sepPlots + 1, multiplied by two if \code{plotit == "both"}.  If only one device is given, all the plots will
go to this device. If none is given, all plots will go to the active device}

\item{sepPlots}{List of groups of indexes that should be plotted for
a limited group of descriptors in separate devices (normally pdfs).
It can either be in a
format similar to \code{sepPlots = list(p1 = c("s1_coh6", "s1_coh6_VH"), 
                p2 = c("NDVI", "BSI"))}, which will plot to two extra devices,
two s1-signals to the first and 2 s2-indexes to the second, for the
default descriptor \code{P50}. Otherwise it can be given in the format
sepPlots = list(bands = list(p1 = c("s1_coh6", "s1_coh6_VH"),
p2 = c("NDVI", "BSI"), stats = c("P50", "MEAN"), which
plot to the same extra devices, but will also include the \code{MEAN} values.
The number of groups in \code{sepPlots} should be reflected by
the number of devices \code{devs}}

\item{zeroshift}{Logical; should all activity centred time series be shifted according to the value at week0?}

\item{xlab}{A title for the x-axis}

\item{ylab}{A title for the y-axis}

\item{main}{A title for the plot (the statistical descriptor and possibly the value of dels)}

\item{yrange}{Plotting range for the y-axis}

\item{FOIinfo}{\code{\link{data.frame}} with information about activities and the corresponding FOIs. See details}

\item{ind}{The name of the index used in a particular calcuation}

\item{EVDcol}{The column giving the activity date}

\item{legend}{If legend is TRUE or "YES", a legend will be added to each plot.
If legend = "BEFORE"),  there will be a separate page at the start of the
document with all legends instead of a legend on each plot.
The argument is not case sensitive.}

\item{addNum}{Logical; should event number types be added to plots}

\item{iprint}{indicates whether the routine should print some information (0 = nothing, 1 some)}
}
\value{
A list of list of means, standard deviations and the y-limits used
in the plots (can be modified for later call of the function).
The list is organized as
\code{eventMeans[[statistic]][[iidel]]} where the statistic is the name
of the statistic for which a mean and std has been computed, and iidel
refers to which number of \code{dels}  we are looking at. If \code{dels = 0},
\code{iidel = 1}.
}
\description{
The function computes mean and standard deviation for a particular
activity type, for the statistics included in the stats variable.
}
\details{
If the means should only be computed for a particular activity
type, the different activity types have to be in the column
\code{dat$EVENT_TYPE}. The FOI IDs need to be in a column with the name
\code{dat$ID}. The column names of the statistics in the data.frame
need to match the list of statistics listed in \code{stats}. \cr \cr
The \code{dat data.frame} should include all the activity-centred weekly
observations for all FOIs of interest. The \code{data.frame} needs data with
particular column names as a minimum (they do not have to be in the same order).
The column names are not case sensitive.
\itemize{
\item{ID}{ - The IDs of the FOIs}
\item{EVENT_TYPE}{ - This column gives a code for activity type, and should
match the activities in the \code{events} argument if the function
should only operate on the activity types of that argument.}
\item{WEEK}{ - The week number of the observation relative to an activity (\code{week0}).}
\item{DOY}{ - Which day of the year the activity happened. This is necessary to separate
different activities at the same FOI}
\item{COUNT}{ - The number of acquisitions per week, only used if
\code{weighted = TRUE}}
\item{MEAN, MIN, etc}{ - The data frame must include columns with names matching the
statistic descriptors in the argument \code{stats} for which calculation
is requested. \code{stats} that are not matched by a column name in \code{dat} will be
ignored. }
}
}
\references{
{
\insertRef{Zielinski2022}{SALMS}
}
}
