% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signalProbability.R
\name{signalProbability}
\alias{signalProbability}
\title{Probability plots for signals and descriptors}
\usage{
signalProbability(
  evres,
  dfs,
  weeks = 1:15,
  week0 = 7,
  selwks = 9:12,
  selstats = "P50",
  stats,
  allstats,
  weekcorrect = 0,
  events,
  inds,
  cases,
  plotit = TRUE,
  devs,
  sortedRasters = TRUE,
  computeAll = FALSE,
  drop = TRUE
)
}
\arguments{
\item{evres}{A 5-dimensional array with means and probabilities, generated by functions in
the SALMS-package. See below for more details}

\item{dfs}{a list with significance values, the output of \\code{\\link{signalTTest}}}

\item{weeks}{The number of weeks used for the signal, taken as all values of \code{dat$WEEK} if missing}

\item{week0}{If a signal consists of \code{weeks}, \code{w0} describes the week to compare with}

\item{selwks}{For which weeks should the descriptive statistics be shown}

\item{selstats}{For which descriptors should the time series of values and probabilities be shown}

\item{stats}{The statistical descriptors}

\item{allstats}{If the argument \code{stats} includes fewer descriptors than the ones
that were used for creating \code{evres}, it is recommended to include
\code{allstats} with the names of the descriptors used for \code{evres}.
If not given, the function will assume that the names are the same as in
\code{dfs}, and that their order is the same as for the default descriptors
of the package. If these assumptions are correct, the argument is not necessary.}

\item{weekcorrect}{If the week numbers should be shifted, so that 0 will be one or more weeks after week0}

\item{events}{A vector with the events under \code{dat$EVENT_TYPE} that should be included}

\item{inds}{Array with names of all indices, same order as used in evres}

\item{cases}{Array with the number of cases for each combination index and descriptor}

\item{plotit}{Should plots be generated? If \code{TRUE} or \code{"both"}, all plots
will be generated. If \code{plotit = FALSE}, no plots will be generated.
Otherwise it can either be \code{"graph"} or \code{"raster"} for the two different
plot types (see details)}

\item{devs}{An optional parameter for plotting devices. It is only necessary if the user
would like to split graphs and rasters in two different documents, otherwise
all output will go to the active device, which can be a pdf}

\item{sortedRasters}{Logical - if TRUE, there will be a range of plots that are sorted
for each descriptor/week, subsequently}

\item{computeAll}{Show the probablities also for indexes where none of the descriptors
are estimated to give a change significantly different from zero.}

\item{drop}{Logical; if TRUE and there is only one activity type, the first
dimension (represting activity type) will be dropped from the \code{evres}-variable.
The variable will then be 4-dimensional instead of 5-dimensional.}
}
\value{
The function will return the evres array, updated with the probabilities. The array will have
5 dimensions (four if there is only one activity type and \code{drop = TRUE}). See also details above
}
\description{
The function plots as a map the probability of being above zero or above
a simple threshold for different
signals and descriptors
}
\details{
The function will estimate the average probabilities of the post-activity
signal being above a threshold for a set of FOIs (features/areas of interest).

The statistical descriptors do not have to be the same as in the previous functions.
However, it would be the recommendation to run a set analyses with the same selection.
Changing the selection could be a source of errors. There is some functionality to
match different selections of descriptors, but this is not fail proof.

The function can produce two different types of plots, in addition to returning the results.
It is possible to direct these plots to one or two devices, according to the value of \code{devs}

The first type of plots are the probability raster maps, which are plotted if \code{plotit} is \code{TRUE},
\code{"both"} or \code{"raster"}. The first set of these rasters will
have the descriptors on the x-axis and the different
indexes on the y-axis. The second set will have the weeks on the x-axis, for a particular choice of
descriptors.

The probability maps, particularly the first set, give an indication of which signal and
which descriptor is the most suitable for a
group of FOIs. A set of plots are produced. First of all, the plots show different values for
p > 0 and p > M, where M is a threshold value that can be used to identify an activity.

For the first set of plots with descriptors on the x-axis, there will be separate plots for
each week of the \code{selwks} parameter. The week number is
the number after "W" in the heading.
If there are more than one group of activity types, these
are given after "G". If there is only one activity type, this part of the header is dropped.

The next number is the highest probability in this map. The number after P indicates the
maximum number of FOIs/activities that have been used for this raster. The last part
depends if \code{sortedRasters = TRUE}. If yes, the function will also generate a range
of plots where the rasters are subsequently ordered according to the values of each
individual descriptor/week. Particularly for the descriptors, this gives the opportunity
to compare the different descriptors with each other. The sorted rasters can be identified
with "S.ind" as a part of the header, followed by the sorting variable.

The second group of plots are the rasters, which are plotted if \code{plotit} is \code{TRUE},
\code{"both"} or \code{"graphs"}
The function will also produce a set of graphs of the change of signal values and the
probabilities. These might be easier to interpret if \code{computeAll = TRUE}.

evres is a 5-dimensional array, with information about the mean changes
and probabilities of the change
being above a threshold. It can both be an input argument and the result of different functions in the
SALMS-package. If there is only one activity type, the first dimension can be dropped.
The dimensions are \code{c(events, inds, stats, evres-var , weeks)}. The content of the
array is defined by the different values of \code{evres-var}. These can be:
\enumerate{
\item{1 \code{mean(x)}}
\item{2 \code{cumsum(x)}}
\item{3 \code{std(x)}}
\item{4 \code{std(cumsum(x))}}
\item{5 \code{mean(x)*sign(mean(xx))} where \code{xx} is the signal change for the weeks \code{selwks}  (Making sure the average change after an activity is positive)}
\item{6 \code{cumsum(x)*sign(mean(xx))} where \code{xx} is the signal change for the weeks \code{selwks}  (Making sure the average change after an activity is positive)}
\item{7 \code{p0} - The probability of the change being larger than 0}
\item{8 \code{p01} - The probability of the change being later than a threshold}
\item{9 \code{p0c} - The probability of the cumulative change being larger than 0}
\item{10 \code{p01c} - The probability of the cumulative change being larger than a threshold}
\item{11 \code{max(po)} - The highest p0 of interesting weeks}
\item{12 \code{max(p01)} - The highest p01 of interesting weeks}
\item{13 The value of the change relative to highest}
\item{14 The value of the change relative to lowest}
\item{15 The value \code{p0} relative to the highest}
\item{16 The value \code{p0} relative to the lowest}
}
}
\examples{
ddir = system.file("extdata", package = "SALMS")
odir = paste0(ddir, "/signals")

hdir = tempdir()
# Read all signals and find the names of all signals/indicators
allvar = readSignals(odir)
inds = names(allvar)
# The function first needs the output from allSignalMeans and allSignalTTests
# Store all the results from the calculations in evress, which contains an 
# array with the means and standard deviations
evress = allSignalMeans(allvar, events = 1, first = FALSE, 
                        dels = 0, weighted = FALSE, inds = inds, weeks = 1:15, week0 = 7, 
                        plotit = FALSE, zeroshift = TRUE, iprint = 1)
evres = evress$evres
cases = evress$cases

devs = list()
dfs = allSignalTTests(allvar, first = FALSE, dels = 0, weighted = FALSE,
                     inds = inds, weeks = 1:15, 
                     week0 = 7, plotit = FALSE) 

pdf(paste0(hdir, "/sigProb.pdf"))
evres1 = signalProbability(evress$evres, dfs, week0 = 7, weeks = 1:15, 
    selwks = 9:13, selstats = "P50", weekcorrect = 0, 
    events = 1, inds = inds, cases = cases)
dev.off()


}
\references{
{
\insertRef{Zielinski2022}{SALMS}
}
}
