% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signalTTest.R
\name{signalTTest}
\alias{signalTTest}
\title{Run a t-test of the differences between week 0 and the other weeks}
\usage{
signalTTest(
  dat,
  stats,
  events,
  first = FALSE,
  dels = 0,
  weighted = FALSE,
  weeks,
  week0 = 6,
  weekcorrect = 0,
  plotit = FALSE,
  devs,
  sepPlots = NULL,
  main,
  ind = "",
  yrange = NULL,
  legend = TRUE
)
}
\arguments{
\item{dat}{A data.frame with FOI activity data. See description below for
more information about the content.}

\item{stats}{A vector with the statistic descriptors of \code{dat},
for which means should be computed.}

\item{events}{A vector with the events under \code{dat$EVENT_TYPE} that
should be included}

\item{first}{Boolean - TRUE if only the first event of the year of each
FOI should be included.}

\item{dels}{a vector of minimum number of days between two activities for
one of them to be included in the mean.}

\item{weighted}{Boolean - TRUE if the mean should be weighted according
to the number of acquisitions used for the weekly mean.}

\item{weeks}{The number of weeks used for the signal, taken as all values of
\code{dat$WEEK} if missing. Only necessary if \code{plotit = TRUE}}

\item{week0}{If a signal consists of \code{weeks}, \code{w0} describes
the week of the activity}

\item{weekcorrect}{If the week numbers should be shifted, so that 0 will be one or more weeks after week0}

\item{plotit}{Boolean - should the function plot the result of the t-tests?}

\item{devs}{One or more devices for plotting the results of the function,
or one or more file names that will be opened as pdfs in the function.
The number of devices should reflect the number of groups in sepPlots + 1.  If only one device is given, all the plots will
go to this device. If none is given, all plots will go to the active device}

\item{sepPlots}{List of groups of indexes that should be plotted for
a limited group of descriptors in separate devices (normally pdfs).
It can either be in a
format similar to \code{sepPlots = list(p1 = c("s1_coh6", "s1_coh6_VH"), 
                p2 = c("NDVI", "BSI"))}, which will plot to two extra devices,
two s1-signals to the first and 2 s2-indexes to the second, for the
default descriptor \code{P50}. Otherwise it can be given in the format
sepPlots = list(bands = list(p1 = c("s1_coh6", "s1_coh6_VH"),
p2 = c("NDVI", "BSI"), stats = c("P50", "MEAN"), which
plot to the same extra devices, but will also include the \code{MEAN} values.
The number of groups in \code{sepPlots} should be reflected by
the number of devices \code{devs}}

\item{main}{The first part of the title of the plots (the statistic and the
number of days without other activities will be added to the title)}

\item{ind}{The name of the index used in a particular calcuation}

\item{yrange}{Plotting range for the y-axis}

\item{legend}{If legend is TRUE or "YES", a legend will be added to each plot.
If legend = "BEFORE"),  there will be a separate page at the start of the
document with all legends instead of a legend on each plot.
The argument is not case sensitive.}
}
\value{
A list of list of means, standard deviations and the number of
FOIs included in each computation.  The list is organized as
\code{eventMeans[[statistic]][[iidel]]} where the statistic is the name
of the statistic for which a mean and std has been computed, and iidel
refers to which number of \code{dels}  we are looking at. If \code{dels = 0},
\code{iidel = 1}
}
\description{
The function runs a comparison of the indicator values between week 0 and the weeks before and after
an activity.
}
\details{
The \code{dat} data.frame needs data with particular column names for this
function. The column names are not case sensitive, and the underscores
in the list below can be omitted.
\describe{
\item{EVENT_TYPE}{This column is necessary and should include activities
matching the activities in the \code{events} argument if the function
should only operate on the activity types of that argument.}
\item{ID}{The ID of the FOIs}
\item{MEAN}{The data frame must include columns with names matching the
statistic descriptors in the argument \code{stats} for which calculation
is requested.  }
\item{stats}{that is not matched by a column name in \code{dat} will be
ignored. }
\item{WEEK}{The week number of the observation relative to an activity.
It is not necessary that the activity happens at week 0, only that all
activities happen in the same week.}
\item{DOY}{Which day of the year the activity happened}
\item{COUNT}{The number of acquisitions per week, only used if
\code{weighted = TRUE}}
}

If the means should only be computed for a particular activity
type, the different activity types have to be in the column
\code{dat$EVENT_TYPE}. The FOI IDs need to be in a column with the name
\code{dat$ID}. The column names of the statistics in the data.frame
need to match the list of statistics listed in \code{stats}.

The argument \code{sepPlots} is a way of reducing the number of pages per
pdf. As an example, we could set
\code{sepPlots = list(p1 = c("s1_coh6", "s1_coh6_VH"), p2 = c("NDVI", "BSI"))}.
The function will then plot the two different coherence plots to devs[2] and
the two S2-indexes to devs[3] (devs[1] is for all output).
}
